Dashboard: share the series helpers via theme; labels state the retained span, never 'whole run'
