+++ src/traceml_ai/aggregator/display_drivers/nicegui_sections/theme.py
@@ -20,6 +20,7 @@ the per-section modules construct their UI with the CSS classes defined here.
 from __future__ import annotations
 
 import os
+from datetime import datetime
 from typing import Any, Dict, List, Optional, Tuple
 
 # --- Brand chrome tokens (DESIGN.md) -------------------------------------
@@ -649,3 +650,166 @@ def single_line_options(
             }
         ],
     }
+
+
+# Absent-value marker, shared by every block so one page never spells
+# absence two ways.
+NA = "n/a"
+
+
+# --- shared series/format helpers (used by System and Process) ------------
+
+
+def format_window(window_s: float) -> str:
+    """The rolling window in words: '30 s', '2 min'.
+
+    The payload picks round windows (see ``choose_window_s``) so this reads
+    as a duration a person recognises.
+    """
+    if not window_s or window_s <= 0:
+        return ""
+    if window_s < 60:
+        return f"{window_s:.0f} s"
+    return f"{window_s / 60.0:.0f} min"
+
+
+def format_span(seconds: Optional[float]) -> str:
+    """The window a chart covers, as one phrase in its label: 'last 3 min'."""
+    if not seconds or seconds <= 0:
+        return ""
+    if seconds < 90:
+        return f"last {seconds:.0f} s"
+    return f"last {seconds / 60.0:.0f} min"
+
+
+def format_gb_pair(used_bytes: Any, total_bytes: Any) -> Tuple[str, str]:
+    """A level against its capacity: ('6.3', '/ 16.1 GB')."""
+    used = gb(used_bytes) if used_bytes is not None else None
+    if used is None:
+        return (NA, "")
+    num = f"{used:.1f}"
+    total = gb(total_bytes) if total_bytes is not None else None
+    if total is None or total <= 0:
+        return (num, "GB")
+    total_s = f"{total:.0f}" if total >= 100 else f"{total:.1f}"
+    return (num, f"/ {total_s} GB")
+
+
+def cpu_axis_max(values: List[Any]) -> float:
+    """Zero-anchored ceiling whose half is a whole percent (0 / 5 / 10)."""
+    vals = [float(v) for v in values if v is not None]
+    peak = (max(vals) * 1.2) if vals else 0.0
+    for top in (4.0, 10.0, 20.0, 30.0, 40.0, 60.0, 80.0, 100.0):
+        if peak <= top:
+            return top
+    return 100.0
+
+
+def sparkline_svg(
+    values: List[Optional[float]],
+    color: str,
+    *,
+    width: int = 64,
+    height: int = 14,
+) -> str:
+    """Inline SVG polyline; gaps are dropped, an empty trace is no SVG."""
+    points = [(i, float(v)) for i, v in enumerate(values) if v is not None]
+    if not points:
+        return ""
+    lo = min(v for _, v in points)
+    hi = max(v for _, v in points)
+    span = (hi - lo) or 1.0
+    step = width / max(len(values) - 1, 1)
+    inner = height - 4
+    coords = " ".join(
+        f"{i * step:.1f},{1 + inner - (v - lo) / span * inner:.1f}"
+        for i, v in points
+    )
+    return (
+        f'<svg viewBox="0 0 {width} {height}" '
+        f'style="width:{width}px;height:{height}px;vertical-align:middle">'
+        f'<polyline points="{coords}" fill="none" stroke="{color}" '
+        'stroke-width="1.4"/></svg>'
+    )
+
+
+def num(value: Any, fmt: str = "{:.0f}") -> str:
+    return fmt.format(float(value)) if value is not None else NA
+
+
+def shared_run_axis(
+    run_t: List[float], prun: List[Dict[str, Any]]
+) -> Optional[Tuple[float, float]]:
+    """One (anchor, span) covering both whole-run series.
+
+    Returns the newest clock across the two and the reach back to the
+    oldest, so both charts can be pinned to the same interval.
+    """
+    starts = [run_t[0]] + [e["t"][0] for e in prun if e.get("t")]
+    ends = [run_t[-1]] + [e["t"][-1] for e in prun if e.get("t")]
+    if not starts or not ends:
+        return None
+    newest = max(ends)
+    return (newest, max(newest - min(starts), 1.0))
+
+
+def newest_epoch(x_time: List[str]) -> Optional[float]:
+    """Epoch seconds of the newest parseable stamp, for the clock axis."""
+    for value in reversed(x_time):
+        try:
+            return datetime.fromisoformat(value).timestamp()
+        except Exception:
+            continue
+    return None
+
+
+def relative_seconds(x_time: List[str]) -> List[Optional[float]]:
+    """Seconds before the newest sample (<= 0), aligned with ``x_time``."""
+    stamps: List[Optional[float]] = []
+    for s in x_time:
+        try:
+            stamps.append(datetime.fromisoformat(s).timestamp())
+        except Exception:
+            stamps.append(None)
+    present = [t for t in stamps if t is not None]
+    if not present:
+        return []
+    last = max(present)
+    return [t - last if t is not None else None for t in stamps]
+
+
+def apply_span_axis(
+    options: Dict[str, Any], span: float, newest_epoch: Optional[float] = None
+) -> None:
+    """Pin a chart to its span and label it in wall-clock time.
+
+    The x values are seconds before the newest sample, which keeps the
+    series arithmetic simple, but a reader debugging a slowdown needs the
+    clock: it is what their logs and every other dashboard are keyed on,
+    and it is what the Process block beside this one already shows. The
+    formatters convert on the fly from the newest sample's epoch, and the
+    hover label carries both readings ("19:10 · 45 min ago") so the axis
+    and the tooltip never speak two different vocabularies.
+    """
+    span = max(float(span), 1.0)
+    axis = options["xAxis"]
+    axis["min"] = -span
+    axis["max"] = 0
+    axis["interval"] = span / 3.0
+    if newest_epoch is None:
+        axis["axisLabel"]["show"] = False
+        return
+    axis["axisLabel"]["show"] = True
+    clock = (
+        "const d=new Date((%f+%%s)*1000);const q=n=>('0'+n).slice(-2);"
+        "const c=q(d.getHours())+':'+q(d.getMinutes())%s;"
+        % (float(newest_epoch), "+':'+q(d.getSeconds())" if span < 600 else "")
+    )
+    axis["axisLabel"][":formatter"] = "v=>{%s return c;}" % (clock % "v")
+    pointer = options.get("tooltip", {}).get("axisPointer", {}).get("label")
+    if pointer is not None:
+        pointer[":formatter"] = (
+            "p=>{%s const s=Math.round(-p.value);"
+            "return c+(s<1?' · now':(s<120?' · '+s+' s ago':"
+            "' · '+Math.floor(s/60)+' min ago'));}" % (clock % "p.value")
+        )
